+++ api_v2/views/__init__.py
@@ -2,12 +2,17 @@
 
 from .views import *
 
+from .background import BackgroundFilterSet, BackgroundViewSet
+
 from .creature import CreatureFilterSet, CreatureViewSet
 from .creature import CreatureTypeViewSet
-
-from .background import BackgroundFilterSet, BackgroundViewSet
 
 from .document import DocumentViewSet
 from .document import RulesetViewSet
 from .document import PublisherViewSet
 from .document import LicenseViewSet
+
+from .item import ItemFilterSet, ItemViewSet
+from .item import ItemSetFilterSet, ItemSetViewSet
+from .item import ArmorFilterSet, ArmorViewSet
+from .item import WeaponFilterSet, WeaponViewSet
